+++ CHANGELOG.md
@@ -2,8 +2,14 @@
 
 ## [unreleased]
 
+- Change the syntax of plugin actions from `plugin/action` to `plugin.action` for better compatibility with other tools.
+
+### Internal
+
+- Improve the robustness of the publish pipeline by migrating Bash and Python scripts to `cargo xtask`.
 
 ## [0.51] 2024-02-18
+
 ## Added
 
 - Input history of providers are now persistent.